Rename StatsBuffer to TimedStore in utils
